+++ tests/File/Content/AtPathTest.php
@@ -139,7 +139,7 @@ class AtPathTest extends TestCase
                 $empty->foreach(static function() use (&$called) {
                     ++$called;
                 });
-                $this->assertSame(0, $called);
+                $this->assertSame(1, $called);
 
                 $called = 0;
                 $extra->foreach(static function() use (&$called) {