+++ tests/File/Content/ChunksTest.php
@@ -108,7 +108,7 @@ class ChunksTest extends TestCase
                 $empty->foreach(static function() use (&$called) {
                     ++$called;
                 });
-                $this->assertSame(0, $called);
+                $this->assertSame(1, $called);
 
                 $called = 0;
                 $extra->foreach(static function() use (&$called) {
@@ -139,7 +139,7 @@ class ChunksTest extends TestCase
                 $shouldBeEmpty->foreach(static function() use (&$called) {
                     ++$called;
                 });
-                $this->assertSame(0, $called);
+                $this->assertSame(1, $called);
 
                 $called = 0;
                 $shouldBeTheSame->foreach(static function() use (&$called) {