+++ tests/File/Content/NoneTest.php
@@ -69,7 +69,7 @@ class NoneTest extends TestCase
         $shouldBeEmpty->foreach(static function() use (&$called) {
             ++$called;
         });
-        $this->assertSame(0, $called);
+        $this->assertSame(1, $called);
 
         $called = 0;
         $shouldBeTheSame->foreach(static function() use (&$called) {