+++ tests/File/Content/OfStreamTest.php
@@ -115,7 +115,7 @@ class OfStreamTest extends TestCase
                 $empty->foreach(static function() use (&$called) {
                     ++$called;
                 });
-                $this->assertSame(0, $called);
+                $this->assertSame(1, $called);
 
                 $called = 0;
                 $extra->foreach(static function() use (&$called) {